Correctly build static binaries in parallel
I've misunderstood how .NOTPARALLEL worked, and this caused everything
that should be done in parallel to not be run in parallel, and the
things that not should be done in parallel, be built in parallel.

By reworking the ninja call in to a grouped target, and just expanding
it, it will be called only once, and ninja will be told to ensure all
targets we care about exists.